Add inline match details to player history
